fix(file-input): validate accept entries using suffix matching (#2014)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

Currently the file input validators have a regex that take all dots from
the file name and taking it as the extension, for example:
"file-name.0.0.1.zip" would return extension to be ".0.01.zip" while we
have `accept=".zip"` and that would fail.

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: CDE-2971

## What is the new behavior?

Use suffix matching by each entry in accept so we can compare `.zip` to
the file name and dynamically take only the last part of the file name.

## Does this PR introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

---------

Co-authored-by: Valentin Mladenov <valentin.mladenov@broadcom.com>