fix(datagrid): remove overflow from column separator and select sr label (#2053) (backport to main) (#2054)
Backport of #2053 

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

There is a button in the column separator which have 35px height and
control labels in checkboxes which have 20/24px height plus padding
which creates overflow and scroll in some scenarios which is not visible
but trough drag and drop functionality implemented on some places it can
be scrolled.

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: N/A

Column separator button height is set to fit the container and control
labels which are used for screen reader only are set to have 1px width
and height.

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

---------

Co-authored-by: GitHub <noreply@github.com>